Quite obviously the parsing should check for integers, not bytes.
git-svn-id: svn+ssh://src.earth.threerings.net/narya/trunk@3353 542714f4-19e9-0310-aa3c-eee0fc999fb1

+++ src/java/com/threerings/media/image/ColorPository.java
@@ -102,7 +102,7 @@ public class ColorPository implements Serializable
         {
             // Check if the string is itself a number
             try {
-                byte id = Byte.parseByte(name);
+                int id = Integer.parseInt(name);
                 if (colors.containsKey(id)) {
                     return id;
                 }
